Include 'useSingleRequestParameter=static' for better ergonomics in examples

+++ typescript/vegobjekter/schemas/adresseSchema.ts
@@ -2,5 +2,5 @@ import { z } from 'zod'
 
 export const adresseSchema = z.object({
   navn: z.string(),
-  adressekode: z.number(),
+  adressekode: z.number().int(),
 })
