+++ typescript/vegobjekter/schemas/kryssystemSchema.ts
@@ -3,8 +3,8 @@ import { trafikantgruppeSchema } from './trafikantgruppeSchema'
 import { z } from 'zod'
 
 export const kryssystemSchema = z.object({
-  kryssystem: z.number(),
-  kryssdel: z.number(),
+  kryssystem: z.number().int(),
+  kryssdel: z.number().int(),
   retning: z.lazy(() => retningSchema),
   trafikantgruppe: z.lazy(() => trafikantgruppeSchema),
   meter: z.number().optional(),
