+++ typescript/vegobjekter/schemas/riksvegruteSchema.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod'
 
 export const riksvegruteSchema = z.object({
-  enumid: z.number(),
+  enumid: z.number().int(),
   riksvegrute: z.string(),
 })
