+++ typescript/vegobjekter/schemas/sideanleggSchema.ts
@@ -3,8 +3,8 @@ import { trafikantgruppeSchema } from './trafikantgruppeSchema'
 import { z } from 'zod'
 
 export const sideanleggSchema = z.object({
-  sideanlegg: z.number(),
-  sideanleggsdel: z.number(),
+  sideanlegg: z.number().int(),
+  sideanleggsdel: z.number().int(),
   retning: z.lazy(() => retningSchema),
   trafikantgruppe: z.lazy(() => trafikantgruppeSchema),
   meter: z.number().optional(),
